Optimized SIMD version with float calculations. Removed/commented stray outputs.
If the force is high, the minimization step size lambda can be low. By allowing for a really large LAMBDATOL (1e-4), the system was not able to minimize effectively.
To overcome this issue, I have created a function which sets the floor of LAMBDATOL based on the maximum force in the system.